+++ xelis_daemon/src/core/storage/mod.rs
@@ -22,7 +22,7 @@ pub type Tips = HashSet<Hash>;
 #[async_trait]
 pub trait DifficultyProvider {
     async fn get_height_for_block_hash(&self, hash: &Hash) -> Result<u64, BlockchainError>;
-    async fn get_timestamp_for_block_hash(&self, hash: &Hash) -> Result<u128, BlockchainError>;
+    async fn get_timestamp_for_block_hash(&self, hash: &Hash) -> Result<u64, BlockchainError>;
     async fn get_difficulty_for_block_hash(&self, hash: &Hash) -> Result<Difficulty, BlockchainError>;
     async fn get_cumulative_difficulty_for_block_hash(&self, hash: &Hash) -> Result<CumulativeDifficulty, BlockchainError>;
     async fn get_past_blocks_for_block_hash(&self, hash: &Hash) -> Result<Immutable<IndexSet<Hash>>, BlockchainError>;
